factor out slist_to_buffer()
Move some code from `config/slist.c` to `mutt/slist.c` so it can be shared.
